chatbot/requirements.txt
# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# --- Configuration ---
pydantic-settings>=2.4.0
python-dotenv==1.0.1

# --- LLM + Embeddings ---
openai==1.51.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# --- Vector Store ---
faiss-cpu==1.8.0

# --- Document Processing ---
pypdf==4.3.1
beautifulsoup4==4.12.3
tiktoken==0.7.0
python-docx>=1.1.0
openpyxl>=3.1.0
pdf2image>=1.17.0
Pillow>=10.0.0

# --- Supabase pgvector ---
psycopg2-binary>=2.9.9
pgvector>=0.3.0

# --- Caching ---
redis==5.0.8

# --- Resilience ---
tenacity>=8.2.0

# --- Testing ---
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0
